Split symbol table test support setters
