+++ src/resolver/symbol_table_test_support.rs
@@ -1,9 +1,12 @@
 use super::*;
 use crate::ast::AstType;
 
+mod aggregate_metadata;
+mod behavior_metadata;
+
 impl SymbolTable {
     #[cfg(test)]
-    fn find_symbol_mut_for_test(
+    pub(super) fn find_symbol_mut_for_test(
         &mut self,
         namespace: Namespace,
         name: &str,
@@ -208,198 +211,6 @@ impl SymbolTable {
     ) {
         if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
             symbol.type_parameter_bound_refs = type_parameter_bound_refs;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_behavior_method_signatures_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        behavior_method_signatures: Option<Vec<MethodSignatureMetadata>>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.behavior_method_signatures = behavior_method_signatures;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_behavior_method_types_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        behavior_method_types: Option<Vec<BehaviorMethodTypeMetadata>>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.behavior_method_types = behavior_method_types;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_behavior_parent_names_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        behavior_parent_names: Option<Vec<String>>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.behavior_parent_names = behavior_parent_names;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_behavior_parent_refs_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        behavior_parent_refs: Option<Vec<BehaviorRefMetadata>>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.behavior_parent_refs = behavior_parent_refs;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_behavior_impl_names_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        behavior_impl_names: Option<Vec<String>>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.behavior_impl_names = behavior_impl_names;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_behavior_impl_refs_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        behavior_impl_refs: Option<Vec<BehaviorRefMetadata>>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.behavior_impl_refs = behavior_impl_refs;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_behavior_required_names_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        behavior_required_names: Option<Vec<String>>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.behavior_required_names = behavior_required_names;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_behavior_required_refs_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        behavior_required_refs: Option<Vec<BehaviorRefMetadata>>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.behavior_required_refs = behavior_required_refs;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_field_count_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        field_count: Option<usize>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.field_count = field_count;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_field_type_names_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        field_type_names: Option<Vec<(String, String)>>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.field_type_names = field_type_names;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_field_types_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        field_types: Option<Vec<(String, AstType)>>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.field_types = field_types;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_variant_names_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        variant_names: Option<Vec<String>>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.variant_names = variant_names;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_variant_owner_name_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        variant_owner_name: Option<String>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.variant_owner_name = variant_owner_name;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_variant_payload_count_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        variant_payload_count: Option<usize>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.variant_payload_count = variant_payload_count;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_variant_payload_type_name_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        variant_payload_type_name: Option<String>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.variant_payload_type_name = variant_payload_type_name;
-        }
-    }
-
-    #[cfg(test)]
-    pub(crate) fn set_variant_payload_type_for_test(
-        &mut self,
-        namespace: Namespace,
-        name: &str,
-        variant_payload_type: Option<AstType>,
-    ) {
-        if let Some(symbol) = self.find_symbol_mut_for_test(namespace, name) {
-            symbol.variant_payload_type = variant_payload_type;
         }
     }
 }
